Move error components to holocene
