+++ src/routes/__layout-root.svelte
@@ -65,7 +65,7 @@
   import Header from './_header.svelte';
   import Notifications from '$lib/components/notifications.svelte';
   import Banners from '$lib/components/banner/banners.svelte';
-  import { ErrorBoundary } from '$lib/components/error-boundary';
+  import { ErrorBoundary } from '$lib/holocene/error-boundary';
   import ScrollToTop from '$lib/holocene/scroll-to-top.svelte';
   import Toaster, { toaster } from '$lib/holocene/toaster.svelte';
